Moved Grab Point struct to KeyPoint.hpp. Moved do_auto_grab to Player.hpp.
