+++ PlayMode.hpp
@@ -8,6 +8,7 @@
 #include "Collisions.hpp"
 #include "DynamicCamera.hpp"
 #include "Player.hpp"
+#include "KeyPoint.hpp"
 
 #include <glm/glm.hpp>
 
@@ -76,22 +77,6 @@ struct PlayMode : Mode {
 		}
 	};
 
-	struct Grab_Point {
-		glm::vec2 position;
-		float past_player_dist;
-		
-		Grab_Point(glm::vec2 position_) : position{position_} {
-			past_player_dist = 0;
-		}
-	};
-
-	struct Spring_Point {
-		glm::vec2 pos;
-		glm::vec2 vel;
-
-		Spring_Point(glm::vec2 pos_, glm::vec2 vel_): pos{pos_}, vel{vel_} {}
-	};
-
 	//----- helper functions ----- see Trello documentation for details
 
 	GLuint load_texture(std::string filename);
@@ -144,8 +129,6 @@ struct PlayMode : Mode {
 		uint8_t downs = 0;
 		uint8_t pressed = 0;
 	} left, right, down, up, space;
-	bool fixed_head = false;
-	bool fixed_tail = false;
 
 	//player struct
 	Player player;
